Fix composite build so that substituted project artifacts are built with a composite context
The same composite context is used to build an artifact for a producer build as for the consumer.
This effectively allows a project to be both a provider and consumer within a single composite
build invocation.

This change exposed an issue with the re-use of 'warm' GradleLauncher instances. This optimization
has been disabled, and projects are now configured once to construct the metadata and again
to build the artifacts.